Flattened Table Data use, other improvements
This update includes an offset script version which accepts a form of
centerline data that has alias names flattened onto it with additional
fields rather than using a separate lookup table. Performs much better.
It also includes a change to the cluster tolerance (to 0 from 2) which
results in improved accuracy of offset points.