Modifications for Easier Debugging
Changed some of the script's layout so that it is easier to debug if
there is anything that it doesn't work well with in the dataset(s) that
are passed into it. Added a single row crash offset script that is
called by the continuousProcessOffsetScript. The single row crash offset
script takes in a single crash at a time and returns a single crash at a
time, with all of the output left up to the continuous process offset
script rather than the old model which had some of the processes in the
called crash offset functions writing directly to the output.